An extra condition exists that can cause the fail
The condition listed if set to SOME and not including the USER (as per edit) is missing.
Cause: One of the following conditions is true: 

The UPN contains an unverified or non-routable domain, such as .local (like joe@contoso.local).
MDM user scope is set to None.
MDM user scope is set to SOME, and excludes the user trying to join.  <---- this is an additional reason I feel, but is missing.